ci: Impl cheap approach for checkstyle9 rev list cooking
If the target branch was forked a long time ago according to the current
state of main branch then it won't be able to find common ancestor and
as a result all N commits of the target branch will be fed to the style
checker. What is not the expected behaviour, but probably it's an
acceptable trade-off.

Where N is the fetch-depth config option used for actions/checkout
GitHub Action and for our Hack step.