ci: Impl simple approach for PR rev list cooking
It tries to fetch more and more with each iteration until it finds
common ancestor (git merge-base).

If main and target branch are not diverted much then one iteration could
be enough. If main is kept up-to-date and target was forked a long time
ago then it needs fetching deeper, esp. the main branch.